sample/fullstack-voice-ai-agent/pyproject.toml
[project]
name = "voice-sample-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.22.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "websockets>=13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# パッケージの指定(appディレクトリ配下を指定)
[tool.hatch.build.targets.wheel]
packages = ["app"]
